+++ packages/conversation/src/models/v1/mod-callback-events/contact-update-event/contact-update-event.ts
@@ -1,10 +1,10 @@
 import { ContactNotification } from '../contact-notification';
-import { ConversationEvent } from '../conversation-event';
+import { ConversationCallbackEvent } from '../conversation-callback-event';
 
 /**
  * This callback is sent when a new contact is updated.
  */
-export interface ContactUpdateEvent extends ConversationEvent{
+export interface ContactUpdateEvent extends ConversationCallbackEvent{
 
   /** Id of the subscribed app. */
   app_id?: string;