+++ packages/conversation/src/models/v1/mod-callback-events/conversation-callback-event/index.ts
@@ -1,0 +1,1 @@
+export { ConversationCallbackEvent } from './conversation-callback-event';