+++ packages/conversation/CHANGELOG.md
@@ -1,3 +1,11 @@
-# Version X.Y.Z
+# Version 0.0.2
 
- - Initial version
+- Initial version. Support for:
+    - Conversation App management
+    - Contact management
+    - Conversation management
+    - Message sending on various channels with various types (including templates v1 and v2) and Message management
+    - Events management
+    - Message transcoding
+    - Capability lookup for a contact
+    - Callback webhooks